patient project/streamlit-app/requirements.txt
 streamlit
 pandas 
 numpy
 sqlalchemy
 confluent-kafka
 plotly
 seaborn
 matplotlib
 psycopg2-binary
